+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/exceptions/AddressTypeException.kt
@@ -1,3 +1,3 @@
-package jp.co.soramitsu.fearless_utils_android.exceptions
+package jp.co.soramitsu.fearless_utils.exceptions
 
 class AddressTypeException : Exception("Address type is wrong")
