+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/ss58/AddressType.kt
@@ -1,0 +1,8 @@
+package jp.co.soramitsu.fearless_utils.ss58
+
+enum class AddressType(
+    val addressByte: Byte
+) {
+    KUSAMA(2.toByte()),
+    POLKADOT(0.toByte())
+}
