+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/ss58/SS58Encoder.kt
@@ -1,10 +1,12 @@
-package jp.co.soramitsu.fearless_utils_android.ss58
+package jp.co.soramitsu.fearless_utils.ss58
 
-import jp.co.soramitsu.fearless_utils_android.exceptions.AddressTypeException
-import jp.co.soramitsu.fearless_utils_android.utils.Base58
+import jp.co.soramitsu.fearless_utils.exceptions.AddressTypeException
+import jp.co.soramitsu.fearless_utils.encrypt.Base58
 import org.spongycastle.jcajce.provider.digest.Blake2b
 
-class SS58Encoder {
+class SS58Encoder(
+    private val base58: Base58
+) {
 
     companion object {
 
@@ -13,8 +15,6 @@ class SS58Encoder {
         private const val PREFIX_SIZE = 2
         private const val PUBLIC_KEY_SIZE = 32
     }
-
-    private val base58 = Base58()
 
     fun encode(publicKey: ByteArray, addressType: AddressType): String {
         val addressTypeByteArray = byteArrayOf(addressType.addressByte)
