+++ fearless-utils/src/test/java/jp/co/soramitsu/fearless_utils/ss58/SS58EncoderTest.kt
@@ -1,15 +1,29 @@
-package jp.co.soramitsu.fearless_utils_android.ss58
+package jp.co.soramitsu.fearless_utils.ss58
 
+import jp.co.soramitsu.fearless_utils.encrypt.Base58
 import org.junit.Test
 
 import org.junit.Assert.assertEquals
+import org.junit.Before
+import org.junit.runner.RunWith
+import org.mockito.Mock
+import org.mockito.junit.MockitoJUnitRunner
 import org.spongycastle.util.encoders.Hex
 
+@RunWith(MockitoJUnitRunner::class)
 class SS58EncoderTest {
+
+    @Mock private lateinit var base58: Base58
+
     private val inputPublicKey = "6addccf0b805e2d0dc445239b800201e1fb6f17f92ef4eaa1516f4d0e2cf1664"
     private val outputAddress = "EzSUv17LNHTU2xdPKLuLkPy7fCD795DZ6d5CnF4x4HSkcb4"
 
-    private val ss58Encoder = SS58Encoder()
+    private lateinit var ss58Encoder: SS58Encoder
+
+    @Before
+    fun setUp() {
+        ss58Encoder = SS58Encoder(base58)
+    }
 
     @Test
     fun encode_called() {
